requirements.txt
python-pptx==0.6.23
pillow>=9.0.0
requests>=2.28.0
aiohttp>=3.8.0
aiofiles>=23.0.0
python-dotenv>=0.19.0
PyMuPDF>=1.23.0
openai>=1.0.0 